Add method to get the address of any capability
Previously it wasn't possible to get the address of MSI and MSI-X
capabilities.

This is necessary for Redox OS to gradually move towards using this
crate for parsing capabilities and in general allows users to parse
capabilties that are unsupported by this crate themself without breaking
when this crate adds support for parsing them itself.